fixed removing unexisting bin folder

+++ run.ps1
@@ -1,15 +1,18 @@
-# Limpar o diretório bin antes de compilar
-Remove-Item -Recurse -Force bin\*
+# Checks if bin folder (folder with compiled files) exists
+if (Test-Path -Path "bin") {
+    # Cleans it
+    Remove-Item -Recurse -Force bin\*
+}
 
-# Compilar arquivos .java para o diretório bin
+# Compiles files into bin folder, if doesn't exist creates it
 javac -d bin -cp bin -sourcepath src $(Get-ChildItem -Recurse -Path src -Filter *.java | ForEach-Object { $_.FullName })
 
-# Verificar se a compilação foi bem-sucedida
+# Checks if previous commands had errors
 if ($?) {
 
     Write-Host "Compilação bem-sucedida!" -BackgroundColor Green -ForegroundColor White
 
-    # Executar o programa Java
+    # Execute java project with args of this script
     java -cp bin pt.iscte.poo.game.Main $args
 } else {
